switch button id to int16: set reserved IDs into the negatives

+++ gfxTabletClient/IDeviceUpdater.cs
@@ -4,6 +4,6 @@
     {
         void ProcessAndEnqueueUpdate(double x, double y, double xmax, double ymax, float pressure, bool inair, bool inverted, bool barrel);
         void Stop();
-        void EmitRawClick(byte button, bool down);
+        void EmitRawClick(short button, bool down);
     }
 }
